AI-Triage-Flow/AI-Triage-Flow/backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-multipart
requests
